Move subsampling routines to dedicated file
